+++ views/index.hbs
@@ -1,0 +1,6 @@
+
+<ul>
+  {{#each tasks}}
+    <li><a href='/tasks/'+{{id}}+'2'>{{id}}. {{body}} {{complete}}</a></li>
+  {{/each}}
+</ul>
