+++ views/layout.hbs
@@ -4,7 +4,7 @@
     <link rel="stylesheet" type="text/css" href="/css/styles.css">
   </head>
   <body>
-    toDoList
+    toDolist
   <br />
   <hr />
    {{{body}}}
